requirements.txt
flask
mongo
dotenv
pillow 
requests 
google-generativeai
langgraph
langchain
langchain-google-genai
pydantic
httpx
langchain-openai
